+++ client/components/navbar.jsx
@@ -1,30 +1,35 @@
 import React from 'react';
-import AppContext from '../lib/app-context';
 
 export default class Navbar extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = ({
+      searchBox: ''
+    });
+    this.handleChange = this.handleChange.bind(this);
+    this.handleSubmit = this.handleSubmit.bind(this);
+  }
 
-  componentDidMount() {
-    fetch(`/api/countries/${this.props.country}`, {
-      method: 'GET',
-      headers: {
-        'Content-Type': 'application/json'
-      }
-    })
-      .then(response => response.json())
-      .then(result => {
-        this.setState({
-          searchBox: '',
-          countries: result
-        });
-      });
+  handleChange(event) {
+    this.setState({
+      searchBox: event.target.value
+    });
+  }
+
+  handleSubmit(event) {
+    event.preventDefault();
+    window.location.hash = `#search-results?country=${this.state.searchBox}`;
+    this.setState({
+      searchBox: ''
+    });
   }
 
   render() {
     return (
       <div className='navbar-container'>
         <div className='search-box-container'>
-          <form onSubmit={this.props.handleSubmit}>
-            <input className="search-box" type="text" value={this.props.searchBox} onChange={this.props.handleChange} name="searchBox" placeholder="search for a country"/>
+          <form onSubmit={this.handleSubmit}>
+            <input className="search-box" type="text" value={this.state.searchBox} onChange={this.handleChange} name="searchBox" placeholder="search for a country"/>
             <button>Submit</button>
           </form>
         </div>
@@ -35,5 +40,3 @@ export default class Navbar extends React.Component {
     );
   }
 }
-
-Navbar.contextType = AppContext;
